Fix gRPC build on macOS
On macOS with xcode 16, compilation of zlib and abseil will fail because of some conflicts with the standard library.
This was fixed by using zlib 1.3.1 in gRPC, and patching abseil.

Furthermore, CMake 4 removed support for CMake < 3.5, and cares still lists 3.1 as minimum required version.
This was fixed by adding -DCMAKE_POLICY_VERSION_MINIMUM=3.5.