Use correct OpenSSL library and other CMake stuff (#78)
Main change is that we now use the OpenSSL library from the dependencies directory during compilation, not the system install. This is done by setting `OPENSSL_ROOT_DIR`.
The `find_package(OpenSSL)` call also sets the `OPENSSL_INCLUDE_DIR` (with OPENSSL in upper-case) correctly so that gRPC finds it.

I also did a bit of cleanup/reorganization along the way:
- I gave the _test_ directory its own CMakeLists.txt. The
`integration_test` executable will still be placed in the project's
binary directory.
- I created separate object libraries for `fivetran_sdk` and `duckdb`
for better separation of concerns.
- I removed unnecessary linked libraries/include directories. Most of
those are already propagated by linking other targets.
- I removed linking `${_REFLECTION}` which I think is never set and I
don't know what this is.
- Set `DEPENDENCIES_DIR` to a default value instead of raising an error.
With this change, one does not have to configure a separate CMake
variable when loading the project in CLion.
- In Makefile, remove the OpenSSL install directory when reinstalling to
avoid conflicts.