Use single DuckDB instance for all requests (#79)
Fivetran spins up a new container for each job/sync. Therefore, the
motherduck_token is the same during the whole lifetime of the MotherDuck
destination container. This allows us to use a single DuckDB instance.

Each `DestinationSdkImpl` holds an instance of `duckdb::DuckDB` which is
initialized the first time a gRPC request comes in (in `get_duckdb`) and
stays alive for the lifetime of `DestinationSdkImpl`. In practice, this
means that we have one `duckdb::DuckDB` instance per container.

During initialization, the following happens:
- `duckdb::DBConfig` is created
- `duckdb::DuckDB` instance is created in single-attach mode
- core_functions extension is loaded; extensions are per
`duckdb::DatabaseInstance`, not per `duckdb::Connection`

Each request handler creates its own logger. Therefore, setting the
`duckdb_id` in the logger happens in `get_connection` even though the
`duckdb_id` stays the same for the lifetime of `DestinationSdkImpl`. The
logger now logs both the duckdb_id and connection_id to make it possible
to distinguish queries from different gRPC requests.

Because the in-memory database `localmem` that is created in
`WriteBatch` and `WriteHistoryBatch` is not deleted now that the
`duckdb::DuckDB` instance stays alive. To start from a clean slate and
prevent issues with concurrent requests, every call to `process_files`
gets its own temporary in-memory database now to create a view using
`duckdb_arrow_scan`.