Add test for parallel DescribeTable requests (#84)
Adds a test for `DescribeTable` calls happening in parallel.

I also sneaked in two other small changes:
- Log the number of columns that the connector is processing in
`DescribeTable`. We have observed crashes before the "Endpoint
<DescribeTable>: ended" log line was printed. This modified log line can
indicate if the crash happened inside the scope of the loop over the
`duckdb_columns` or when/after this scope was left.
- Use `QueryResult::Cast` to get a `MaterializedQueryResult` because
this has better type checking than `unique_ptr_cast`.