Add unittest for `decrypt_stream` (#83)
Adds a test that encrypts and decrypts a random string, then asserts
that the decrypted ciphertext is equal to the plaintext. The
`encrypt_stream` function is located in the test itself because it is
not used anywhere else currently.

The test is part of the `integration_test` executable even though it is
not an integration test, technically speaking. But I think we don't have
enough unit tests yet to split those into separate executables.

The test can be run with `./build/Release/integration_tests "Decrypt is
inverse function of encrypt"`.