Fix Docker build (#91)
- Add `GIT_COMMIT_SHA_OVERRIDE` argument to Docker file. Since #81, git
is required to compute a short commit SHA as version number. However, we
do not copy the .git directory to the Docker build image, hence this
will fail. This adds a new argument to make it possible to pass the SHA
from the outside with `--build-arg GIT_COMMIT_SHA_OVERRIDE=$(git
rev-parse --short HEAD)`.
- Skip adding the _test_ subdirectory in CMake if it is not present.
This is also because we don't copy the test directory to the Docker
build image.
- Force creating a build system based on Makefiles when building gRPC
and Catch2 because we run `make` (instead of, for example, `cmake
--build`) to start the build process. This can cause problems if Ninja
is the default build system generator (e.g. by setting the
`CMAKE_GENERATOR` environment variable).
- Reorganized the `build_test_dependencies` Make target to follow the
style of the other targets with one command per line and avoiding `cd`
when possible. I also renamed the clone directory from "Catch2" to
"catch2" to make the naming consistent with the other directories.