Build connector with pre-compiled static DuckDB libraries (#93)
We need to make sure that the compilation produces one static binary
without dependencies on external shared libraries like libduckdb.
Therefore, we are refining the approach from #85 and link static
libraries instead.

DuckDB releases `static-libs-<os>-<arch>.zip` artifacts that inflate to
a bunch of static libraries that are all required to be linked. I listed
them explicitly in the CMakeLists instead of globbing the directory to
make sure that the libraries are ordered from left to right with the
dependents left of the dependencies. The artifact does not come with
duckdb.hpp, so we copy this header from the `libduckdb-src` artifact
instead.

The `osx` artifacts are platform-specific. `uname -m` on OSX returns
either `x86_64` or `arm64` in the "normal" cases, hence we have to add
handling for the `arm64` case when figuring out the machine
architecture.

The idea of this PR is to unblock the release. I will try using
`FetchContent`/`ExternalProject` which I think will be nicer because we
can still build things ourselves, but heavier than just building
libduckdb.