Set default_collation to default (empty) value on new connections (#96)
To make sure the collation settings are not affected by other
connections and that client and duckling use the same collation for
planning and execution, sets `default_collation` to an empty string for
every new connection. I am rather certain this could have also been done
in `get_duckdb` once, but I'd rather pay the price for a bit more
requests to get extra certainty until we understand the problem better.