Set connection time zone to UTC (#100)
Two tests were failing locally for me because of time zone mismatches. I
think the problem is that `make_timestamp` returns a `TIMESTAMP`, but we
convert Fivetran's `UTC_DATETIME` type to a `TIMESTAMP_TZ`. As a
tactical fix, this sets the time zone to UTC on every connection. I have
to say that it is not 100% clear to me why the tests (e.g. "WriteBatch")
fail.

The problem was introduced when we started to use the pre-built
libduckdb binaries because they come with the ICU extension statically
linked. The problem will go away once we again build libduckdb from
source (without ICU).

Also in this PR:
- Add comment describing why we need to run `SET default_collation`
- Formatting