Replace 'fivetran_test010' database before running integration tests (#101)
We use randomized table and schema names. However, to reduce the chances
of name collisions even more, we start with a fresh database at the
start of tests which we also give a randomized name. The database gets
dropped after the test run.

Furthermore, makes `MD_TOKEN` a global variable that is only read once
from the environment variables, and accepts both `motherduck_token`
(lowercase) and `MOTHERDUCK_TOKEN` (uppercase).