Add class for temporary memory-backed file on Linux (#106)
Adds a `MemoryBackedFile` class which holds an open file descriptor. The
memory-backed files will be used to store the decrypted CSV content in.
The file is unlinked shortly after creation, hence cannot be found in
the file system and is deleted once the file descriptor is released.
Because the file is unlinked, we access it via the path `/dev/fd/N`. The
file descriptor is closed when a `MemoryBackedFile` goes out of scope.
The file is completely in memory on Linux. On macOS, the file lives in
the `/tmp` directory.