Use DuckDB's `read_csv` function for file processing (#107)
In `ProcessFile`, use DuckDB's `read_csv` function instead of reading
from an Arrow table. This reduces the memory footprint of the connector
because we can read the CSV files in batches and stream the results to
MotherDuck's backend.

- `IngestProperties` has a new field `allow_unmodified_string` which is
set to true when reading `UPDATE` files. In these files, any column can
contain a special `"unmodified-string"` value which indicates that the
value doesn't change in the update. This is a string regardless of the
column type, hence we have to read the CSV file with `all_varchar=true`.
The conversion will happen later during the ingestion into the table.
- If the CSV file is decrypted, we use the `decrypt_file` function and
store the decrypted data in a `MemoryBackedFile`. In a follow-up, I will
make sure that the `decrypt_file` writes directly to this file and does
not buffer the data in-between, but this PR is big enough already.
- Because the cursor of the file descriptor referred to via `/dev/fd/N`
on macOS is shared (because it is using `dup`), we have to reset the
cursor to the beginning of the file using `lseek` after every read
operation on macOS. :-/
- We pass the compression type directly to DuckDB's `read_csv` function.
We therefore check if the first four bytes of the file are the zstd
magic bytes.
- When using `read_csv`, we push down the column types to the CSV
reader. We use `column_types` instead of `columns` to not rely on the
order of columns in the CSV file. The column order in the CSV file is
inferred by DuckDB from the header row. We `SELECT` the columns in the
table insertion order at the end of the query.
- We set a bunch of options in the `read_csv` function. I am not sure
whether this makes it too brittle. For example, we enforce `\n` as line
ending. However, given that Fivetran's CSV files are expected to be
well-formed, I think it is fair to make these assumptions.
- We cannot set a timestampformat because we can only set one, but we
might have to deal with native time and UTC time in the same CSV file.
- The CSV block size parameter is now reused to determine the CSV
reader's buffer size. I think we can revisit this later.
- The special null strings that Fivetran sends as part of the
`Write[History]Batch` requests are passed to `read_csv` via its
`nullstr` parameter. DuckDB uses it to recognize `NULL` values even for
column of other types than `VARCHAR`.

A typical query looks like this:
```sql
CREATE VIEW "csv_view" AS
FROM read_csv('/dev/fd/3', auto_detect=true, delim=',', encoding='utf-8', escape='"', header=true, new_line='\n', quote='"', dateformat='%Y-%m-%d', nullstr='magic-nullvalue', allow_quoted_nulls=true, compression='none', column_types={'id':'INTEGER','_fivetran_start':'TIMESTAMP WITH TIME ZONE',})
SELECT "id", "_fivetran_start",
```