Use default max_line_size of 16 MiB for CSV parsing (#110)
TL;DR: Sets the CSV reader options `max_line_size` to 16 MiB and
`buffer_size` to 64 MiB. Removes the CSV block size field from the
configuration form.

Before #107, we were reading CSV files with Arrow. By default, the read
block size was 1 MiB, but it was possible to configure this setting (see
#42). A block in Arrow had to be big enough to contain one CSV line.
After #107, we misused the block size config parameter as buffer size
for DuckDB's CSV reader if it was larger than the default of 32 MB.

Instead of making this a config option and telling users to increase the
block size if they run into issues, we first wanted to implement
automatic retries with ever-increasing max_line_size. However, we
realized that we can simply use a higher default value for
`max_line_size` and hence avoid running into the "Maximum line size
exceeded" error in the first place.

We have to keep in mind that Fivetran can (in the future) submit up to
eight simultaneous `WriteBatch` requests. The memory limit is 1 or 2
GiB. We don't want to push it too far, so we assumed a limit of 512 MiB.
Across eight threads, this means the maximum _buffer_ size is 64 MiB.
The CSV reader should be able to put four lines into the buffer, hence
`max_line_size` is set to 16 MiB.

See the (unused) DuckDB constant `MIN_ROWS_PER_BUFFER = 4`.
https://github.com/duckdb/duckdb/blob/e4fcfd4ddbf9f6acad8dd167124b7dba66c70562/src/include/duckdb/execution/operator/csv_scanner/csv_buffer.hpp#L80

This `max_line_size` is eight times larger than DuckDB's default and 16
times larger than what we used for Arrow. Unless we get error reports,
we can assume that this is more than enough. This also avoids the
complexity of having to implement retries. Note that, if we really need
it, we still have space left. For example, in practice, DuckDB's CSV
reader is fine to put a single line into the buffer. We could also
assume a higher container memory limit.

---------

Co-authored-by: Donny Peeters <46660228+Donnype@users.noreply.github.com>