Use `memfd_create` to create memory-backed file on Linux (#112)
For the memory-backed file (#106), I originally wanted to share as much
code as possible between macOS and Linux. Therefore, I just changed the
file path on Linux to use `/dev/shm` to make the file live in RAM, but
the rest stayed the same.
The problem is that in Docker, `/dev/shm` is only 64 MiB large by
default.
However, want to store files there up to a size of 100 MiB. In Docker, the
tmpfs size can be overwritten with the `shm_size` option, but this would
mean changing the Docker settings on the Fivetran side. `memfd_create`
doesn't have this limitation. Another benefit is that a memfd file is
anonymous immediately without having to call `unlink` first. So I want
to switch to use this instead, even though the code becomes slightly
more platform-specific.